requirements.txt
streamlit==1.31.1
pinecone-client==3.0.2
google-generativeai==0.4.1
sentence-transformers==2.6.1
langchain==0.1.14
langchain-community==0.0.30
pypdf==4.1.0
rank-bm25==0.2.2
python-dotenv==1.0.1
tiktoken